+++ core/src/main/java/testapp/sliubetskyi/core/ui/ILocationTrackerView.java
@@ -1,5 +1,5 @@
 package testapp.sliubetskyi.core.ui;
 
 public interface ILocationTrackerView extends ILocationUpdaterView {
-    void targetDistanceAchieved(float targetDistance);
+    void targetDistanceAchieved(long targetDistance);
 }
